refactor: moved parameters to array up as it is needed by api node
