+++ packages/gatsby-transformer-raml/src/utils/parameters-to-array.js
